+++ src/main/java/com/bs/travelagency/entity/Status.java
@@ -11,7 +11,7 @@ public class Status {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "status_id")
-    private Long id;
+    private int id;
 
     @Column(name = "user_status")
     private String post;
@@ -52,11 +52,13 @@ public class Status {
     //endregion
 
     //region for getter setter
-    public Long getId() {
+
+
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
