+++ src/main/java/com/bs/travelagency/repository/IUserSetupRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository(value = "userSetupRepository")
-public interface IUserSetupRepository extends JpaRepository<User, Long> {
+public interface IUserSetupRepository extends JpaRepository<User, Integer> {
 
     User findByEmail(String email);
 }
